requirements.txt
Flask==0.10.1

alembic==0.8.6

# database
Flask-SQLAlchemy===2.1
SQLAlchemy-Utils==0.32.6
psycopg2==2.6.1

PyYAML==3.11

# apis
Flask-RESTful==0.3.5

# Schema validation
marshmallow==2.7.3
marshmallow-jsonapi==0.7.1
